Created new lib, core_utils and moved a couple of general utils into it
