+++ backend_py/primary/primary/auth/auth_helper.py
@@ -1,5 +1,5 @@
-import os
 import base64
+import os
 import time
 from typing import List, Optional
 
@@ -8,10 +8,10 @@ import msal
 import starsessions
 from fastapi import APIRouter, Request, Response
 from fastapi.responses import RedirectResponse
+from webviz.core_utils.perf_timer import PerfTimer
 
-from primary.services.utils.authenticated_user import AuthenticatedUser
-from primary.services.utils.perf_timer import PerfTimer
 from primary import config
+from primary.services.utils.authenticated_user import AuthenticatedUser
 
 
 class AuthHelper:
