+++ backend_py/primary/primary/routers/observations/router.py
@@ -1,5 +1,4 @@
 import logging
-from typing import List, Optional, Literal
 
 from fastapi import APIRouter, Depends, Query
