+++ backend_py/primary/primary/routers/polygons/router.py
@@ -2,19 +2,17 @@ import logging
 from typing import List, Union
 
 from fastapi import APIRouter, Depends, HTTPException, Query
+from webviz.core_utils.perf_timer import PerfTimer
 
-from primary.services.sumo_access._helpers import SumoCase
-from primary.services.smda_access.stratigraphy_access import StratigraphyAccess
-from primary.services.sumo_access.polygons_access import PolygonsAccess
-from primary.services.smda_access.stratigraphy_utils import sort_stratigraphic_names_by_hierarchy
-from primary.services.smda_access.mocked_drogon_smda_access import _mocked_stratigraphy_access
-from primary.services.utils.authenticated_user import AuthenticatedUser
-from primary.services.utils.perf_timer import PerfTimer
 from primary.auth.auth_helper import AuthHelper
+from primary.services.smda_access.mocked_drogon_smda_access import _mocked_stratigraphy_access
+from primary.services.smda_access.stratigraphy_access import StratigraphyAccess
+from primary.services.smda_access.stratigraphy_utils import sort_stratigraphic_names_by_hierarchy
+from primary.services.sumo_access._helpers import SumoCase
+from primary.services.sumo_access.polygons_access import PolygonsAccess
+from primary.services.utils.authenticated_user import AuthenticatedUser
 
-
-from . import schemas
-from . import converters
+from . import converters, schemas
 
 LOGGER = logging.getLogger(__name__)
