+++ backend_py/primary/primary/services/smda_access/queries/get_well_headers.py
@@ -1,7 +1,7 @@
 from typing import List
 
 
-from primary.services.utils.perf_timer import PerfTimer
+from webviz.core_utils.perf_timer import PerfTimer
 from ..types import WellBoreHeader
 from ._get_request import get
