+++ backend_py/primary/primary/services/sumo_access/polygons_access.py
@@ -6,7 +6,7 @@ import pandas as pd
 import xtgeo
 from fmu.sumo.explorer.objects import PolygonsCollection
 
-from primary.services.utils.perf_timer import PerfTimer
+from webviz.core_utils.perf_timer import PerfTimer
 
 from ._helpers import SumoEnsemble
 from .polygons_types import PolygonsMeta
