+++ backend_py/primary/pyproject.toml
@@ -25,7 +25,7 @@ sumo-wrapper-python = "1.0.6"
 azure-monitor-opentelemetry = "^1.1.0"
 requests-toolbelt = "^1.0.0"
 pottery = "^3.0.0"
-
+core_utils = {path = "../libs/core_utils", develop = true}
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.12.0"
